+++ tine20/Tinebase/EmailUser/Smtp/Postfix.php
@@ -797,7 +797,7 @@ class Tinebase_EmailUser_Smtp_Postfix extends Tinebase_EmailUser_Sql implements 
      */
     protected function _checkDomain($_email, $_throwException = false)
     {
-        return Tinebase_EmailUser::checkDomain($_email, $_throwException, $this->_config['alloweddomains']);
+        return Tinebase_EmailUser::checkAllowedDomain($_email, $_throwException, $this->_config['alloweddomains'], _includeExternalDomains: true);
     }
 
     /**